Changed from Double Wrapper to BigDecimal Wrapper

+++ src/main/java/coj/and/CaloriesCalculator/aliments/AlimentsDto.java
@@ -1,4 +1,6 @@
 package coj.and.CaloriesCalculator.aliments;
 
-public record AlimentsDto(String name, Double calories, Double protein, Double carbs, Double fat, Double fiber) {
+import java.math.BigDecimal;
+
+public record AlimentsDto(String name, BigDecimal calories, BigDecimal protein, BigDecimal carbs, BigDecimal fat, BigDecimal fiber) {
 }
